# docs/a7_openapi.yaml
openapi: 3.0.0
info:
  version: "1.0"
  title: "Eventopia"
  description: "Web Resources Specification (A7) for Eventopia"
servers:
  - url: http://lbaw.fe.up.pt
    description: Production server
...
paths:
...
  /dashboard/future/attendee:
    get:
      operationId: R205
      summary: "R205: Dashboard page"
      description: "Show the dashboard page. Access: USR"
      tags:
        - "M02: Users"

      responses:
        "200":
          description: "Ok. Show the dashboard page."
        "302":
          description: "Redirect after processing the new user information."
          headers:
              Location:
                schema:
                  type: string
                example:
                  302Success:
                    description: "Successful. Redirect to login page."
                    value: "/login"
              
  ######### DELETE #########
...